Simplify MinePlacer interface
The MinePlacer interface was confusing because it allowed for incorrect input eg more mines than the field can take. This simplifies the interface to remove the parameter and uses the field size given in the constructor to determine the maximum amount of mines to place.

This requires updating the unit test to remove the parameter.