Allow game to play in full for the console version
A while loop was implemented in the Game class to allow for a user to
play the game in full. Bugs were found when running all tests. If
intsToReturn is an array of only one int then the call times will always
return 0 because then the calledTimes will equal the
intsToReturn.length .