Add acceptance test for Game class
Created two setUp methods in the Game class with one that can take in an
IRandom instance in order to be able to script the game running. Also
removed the automatic setUp from the Game constructor to avoid errors
related to multiple game set ups.

A test was created for the game.play()
that asserts that when a chosen field is populated and a mine is hit on
the first turn, a test writer is called with the expected revealed
field.